backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.5.0
    # via starlette
astroid==3.2.4
    # via pylint
bcrypt==4.2.0
    # via -r requirements.in
build==1.2.2
    # via pip-tools
click==8.1.7
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   pylint
dill==0.3.8
    # via pylint
ecdsa==0.19.0
    # via python-jose
fastapi==0.115.0
    # via -r requirements.in
greenlet==3.1.0
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
isort==5.13.2
    # via pylint
mccabe==0.7.0
    # via pylint
packaging==24.1
    # via build
passlib==1.7.4
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via pylint
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.5.2
    # via -r requirements.in
pylint==3.2.7
    # via -r requirements.in
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.17
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.35
    # via -r requirements.in
starlette==0.38.5
    # via fastapi
tomlkit==0.13.2
    # via pylint
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.30.6
    # via -r requirements.in
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
